+++ examples/llm/main.py
@@ -10,11 +10,10 @@ from composer.utils import reproducibility
 from omegaconf import OmegaConf as om
 
 from examples.common.builders import (build_algorithm, build_callback,
-                                      build_dataloader, build_logger,
+                                      build_dataloader, build_evaluators, build_logger,
                                       build_optimizer, build_scheduler)
 from examples.common.config_utils import log_config, update_batch_size_info
 from examples.llm.src.model_registry import COMPOSER_MODEL_REGISTRY
-
 
 def build_composer_model(cfg):
     warnings.filterwarnings(
