+++ examples/llm/yamls/evaluation/gpt_neo_eval.yaml
@@ -9,7 +9,7 @@ tokenizer:
 
 model: 
   model_type: pretrained_hf
-  config: EleutherAI/gpt-neo-1.3B
+  config: EleutherAI/gpt-neo-125M
 
 icl_tasks:
   -
@@ -17,7 +17,7 @@ icl_tasks:
     dataset_uri: s3://mosaicml-internal-dataset-lambda/lambada/lambada_test.json
     num_fewshot: 
      - 0
-    batch_size: 16
+    batch_size: 8
     type: language_modeling
     metrics:
       - InContextLearningLMAccuracy
